move JacksonFeatures back to its old package for binary compatibility
