+++ jackson-databind/src/main/java/io/micronaut/jackson/codec/JacksonFeatures.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.jackson.databind;
+package io.micronaut.jackson.codec;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.SerializationFeature;
