+++ jackson-databind/src/main/java/io/micronaut/jackson/codec/JsonMediaTypeCodec.java
@@ -22,7 +22,6 @@ import io.micronaut.context.annotation.Secondary;
 import io.micronaut.core.annotation.Nullable;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.codec.CodecConfiguration;
-import io.micronaut.jackson.databind.JacksonFeatures;
 import io.micronaut.runtime.ApplicationConfiguration;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
